Status_Update

+++ Scripts/Object/Player/Character.cs
@@ -14,10 +14,10 @@ public class Character : Elements
     public int level = 1;
     public int exp = 0;
     public int statusPoint = 0;
-    public int str = 5;
-    public int dex = 5;
-    public int wiz = 5;
-    public int luck = 5;
+    public float str = 5;
+    public float dex = 5;
+    public float wiz = 5;
+    public float luck = 5;
     public bool[] checkEquipItems = { false, false, false, false, false, false, false, false, false };
     public Item[] equipedItems = new Item[9];
 }
